This patch fixes a few classes of compiler warnings and improves the robustness of the code against some string over-runs and out-of-range indices.

+++ libraries/mark6sg/extra/m6sg_blocknrs.c
@@ -1,5 +1,5 @@
 /***************************************************************************
- *   Copyright (C) 2014 by Jan Wagner                                      *
+ *   Copyright (C) 2014-2025 by Jan Wagner                                 *
  *                                                                         *
  *   This program is free software; you can redistribute it and/or modify  *
  *   it under the terms of the GNU General Public License as published by  *
@@ -151,8 +151,8 @@ int main(int argc, char** argv)
             if (bhdr[i].wb_size != fhdr[i].block_size && offset[i] < (file_size[i] - fhdr[i].block_size))
             {
                 fprintf(stderr,
-                    "ERROR: file %d at offset %ld: mismatch in current wb_size %u vs. file block_size %u : blk %u, previous blk %u\n",
-                    i, offset[i], bhdr[i].wb_size, fhdr[i].block_size, bhdr[i].blocknum, blocknum[i]
+                    "ERROR: file %d at offset %ld: mismatch in current wb_size %u vs. file block_size %u : blk %u, previous blk %lu\n",
+                    i, offset[i], bhdr[i].wb_size, fhdr[i].block_size, bhdr[i].blocknum, (unsigned int)(blocknum[i])
                 );
 
                 do
